Generate const-correct cpp iteration when possible
Python-style `for value in sequence` loops over cpp objects compile to
C++ code of the form:

```
iterator it;
// ...
it = sequence.begin();
for(;;) {
  if(!(it != sequence.end())) break;
  value = *it;
  it++;
  // ... (loop body)
}
```

The type of the loop iterator variable is deduced from the return type
of `begin`. Since overlading on const is not currently supported , this
can lead to const-incorrect code when the sequence is const and `begin`
declares a non-const iterator return type.

Cython wrappers for the C++ standard library seem to use the convention
of declaring `const_begin` and `const_end` for the const versions of
`begin` and `end` while specifying that the cname is actually `begin`
and `end`.

This commit tries to use such `const_`-prefixed declarations to deduce
the type of the loop iterator variable when the sequence is const.